feat: use namespace instead of scope for routing
